move methods to service layer from model
